+++ CHANGELOG.md
@@ -4,9 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## v1.3.0 - 2023-09-25
+## v1.3.0 - 2023-09-27
 ### Added
 - `split_and(p: HplPredicate | HplExpression)` function to `hpl.rewrite` module.
+- `simplify(p: HplPredicate | HplExpression)` function to `hpl.rewrite` module.
+- `is_inclusion` and `is_comparison` properties to `BinaryOperatorDefinition`.
 
 ## v1.2.0 - 2023-09-08
 ### Added